move every components under src/components
